etl/requirements.txt
requests==2.32.3
PyYAML==6.0.2
psycopg[binary]==3.2.9
fastapi==0.115.0
uvicorn[standard]==0.30.6
